+++ python/GafferImageUI/ImageReaderUI.py
@@ -213,10 +213,10 @@ Gaffer.Metadata.registerNode(
 
 			"description",
 			"""
-			The colour space of the input image, used to convert the input image to
-			the scene linear space defined by the OpenColorIO config. When set
-			to `Automatic`, the colour space is determined automatically using the
-			function registered with `ImageReader::setDefaultColorSpaceFunction()`.
+			The colour space of the input image, used to convert the input to
+			the working space. When set to `Automatic`, the colour space is
+			determined automatically using the function registered with
+			`ImageReader::setDefaultColorSpaceFunction()`.
 			""",
 
 			"presetNames", OpenColorIOTransformUI.colorSpacePresetNames,
@@ -280,7 +280,7 @@ class _ColorSpacePlugValueWidget( GafferUI.PresetsPlugValueWidget ) :
 			automaticSpace = ""
 			if len( colorSpacePlugs ) and preset == "Automatic" :
 				with IECore.IgnoredExceptions( Gaffer.ProcessException ) :
-					automaticSpace = colorSpacePlugs[0].getValue()
+					automaticSpace = colorSpacePlugs[0].getValue() or "Working Space"
 
 			result.append( {
 				"preset" : preset,